chore(hybridcloud) Remove role switching and update guard method (#52673)
The role switching logic to catch write operations that are missing
outboxes has been causing test flakiness and disrupting development
workflows.

These changes finish replacing role switching with query auditing that
is done during test teardown. Switching to a teardown based audit means
we can't give the precise location of where the bad operation was as the
stack frames have been lost. However, we are able to show the offending
query and its location in the SQL logs for a test.