ref(backup): Use a more exact datetime serializer (#52771)
The default DjangoJSONEncoder inherits from the odd behavior of Python's
own `isoformat` serializer: it includes milliseconds in the output in
all cases EXCEPT those where the value of the milliseconds are `.000`,
in which case it removes them completely. This makes it difficult and
flaky to do simple string comparison between two dates. Instead, a new
wrapper class is introduced to ensure that we always retain the
milliseconds on all date serializations.